backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.11
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
cachetools>=5.3
